chore(remotesessions): address PR review feedback
- Note dcrMaxBodyBytes is 1 MiB in the constant comment.
- Move remoteSession/IssuerView/ClientView builders into internal/mv
  (BuildRemoteSession*View), matching the usersession.go layout.
- Switch AutoRegister and Discover tests to devidptest.Launch instead
  of hand-rolled httptest servers.
- Make DeleteRemoteSessionClient, DeleteRemoteSessionIssuer, and
  RevokeRemoteSession idempotent: pgx.ErrNoRows now returns nil
  (matches DeleteEnvironment/DeleteToolset). Tests updated to assert
  success + no audit entry.
- Fold register_test.go into issuerhandlers_test.go (after Discover).
